Move options to ``sphinx.ext.autodoc._directive_options``
